[#39] Convert to use `Decimal` instead of `Double` for price and timestamp

+++ CryptoPrices/Sources/Data/Sources/NetworkCore/APIModels/APIPrices.swift
@@ -9,5 +9,5 @@ import Foundation
 
 public struct APIPrices: Decodable, Equatable {
 
-    public let prices: [[Double]]
+    public let prices: [[Decimal]]
 }
